+++ lib/decoded_images/image_download_platform/io.dart
@@ -1,0 +1,8 @@
+import 'dart:typed_data';
+
+import 'package:disposable_cached_images/images_isolate.dart';
+
+Future<Uint8List> getImageBytesFromUrl(String url,
+    {Map<String, String>? headers}) {
+  return ImagesIsolate.getNetworkBytes(url, headers);
+}