+++ lib/decoded_images/image_download_platform/stub.dart
@@ -1,0 +1,6 @@
+import 'dart:typed_data';
+
+Future<Uint8List> getImageBytesFromUrl(String url,
+    {Map<String, String>? headers}) {
+  throw UnimplementedError();
+}