+++ lib/image_providers/providers/network.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: curly_braces_in_flow_control_structures
+
 part of disposable_cached_images;
 
 final _networkImageProvider = StateNotifierProvider.autoDispose
@@ -50,7 +52,8 @@ class _NetworkImageProvider extends BaseImageProvider
         }
       }
 
-      ref.read(_usedImageProvider).add(imageInfo);
+      if (providerArguments.keepBytesInMemory)
+        ref.read(_usedImageProvider).add(imageInfo);
     } catch (e, s) {
       onImageError(e, s);
     }